fix: redirect to login on auth error
