+++ app/routes/authenticate.ts
@@ -1,0 +1,50 @@
+import { StatusCodes } from "http-status-codes";
+import { type LoaderFunctionArgs, redirect } from "react-router";
+import { createSession } from "~/.server/services/sessions/createSession";
+import { stytchClient } from "~/.server/vendors/stytch/client";
+import {
+  STYTCH_SESSION_DURATION_MINUTES,
+  STYTCH_SESSION_TOKEN,
+} from "~/config/auth";
+import { appRoutes } from "~/shared/appRoutes";
+import { KEYS } from "~/shared/keys";
+import { ServerError } from "~/types/server";
+
+export async function loader({ request }: LoaderFunctionArgs) {
+  const searchParams = new URL(request.url).searchParams;
+  const token = searchParams.get(KEYS.token);
+  const tokenType = searchParams.get("stytch_token_type");
+
+  try {
+    if (!token) {
+      throw new ServerError(`No token`, StatusCodes.BAD_REQUEST);
+    }
+
+    if (tokenType !== "magic_links") {
+      console.error(`Unsupported token type: '${tokenType}'`);
+      throw new ServerError(
+        `Unsupported token type: '${tokenType}'`,
+        StatusCodes.BAD_REQUEST,
+      );
+    }
+
+    const res = await stytchClient.magicLinks.authenticate({
+      session_duration_minutes: STYTCH_SESSION_DURATION_MINUTES,
+      token,
+    });
+
+    return createSession({
+      key: STYTCH_SESSION_TOKEN,
+      redirectTo: appRoutes("/"),
+      request,
+      token: res.session_token,
+    });
+  } catch (error) {
+    console.error(`Login error: ${error}`);
+    throw redirect(
+      appRoutes("/login", {
+        error: "true",
+      }),
+    );
+  }
+}
